Fix "docker exec" orphan process
When this "dockersh" process is signalled/killed (like when ssh disconnects
due to network issue etc), due to a docker bug, processes that are running
inside 'docker exec' are left as they are i.e., they will continue to
run. This can add up quite a bit of processes over the time. See here for
more information: https://github.com/moby/moby/issues/9098

Until that is fixed, we will handle it here by:

 - Have the pid of exec'ed bash at /tmp/dockersh-exec-${this_process_pid}

 - Install a signal handler here that will send SIGHUP to the pid stored in
   /tmp/dockersh-exec-${this_process_pid}